Big level change
Level system has been changed a lot, now I have an external level editor that wont part of the entry, just to make things easier. (The level editor was hardcoded, so never look inside it.)
A new level was added, but there's not much in it yet. Also the player arrow/text was centered